Improve performance substantially
- Kill off all `coordsToId` stuff and just use rows and cols like a
non-idiot. This simplifies a LOT of the code.
- In `EntitiesContainer` and `GroundsContainer`, shallowly convert
their data to an array and then convert their *row* immutable data
using `immutableToJs`. This ensures that the row data is always the
same array instance until the row is changed which is important for
component purity.
- Use `level` in a few places. Not exactly perf related but more
related to how the code works now.
- Create `TilesRow` component. Decided to keep it in `Tiles` and not
export it; no new file because it’s an implementation detail of
`Tiles`. This improves perf by diffing a much smaller part of the UI.
Rather than doing `shouldComponentUpdate` for thousands of `Tile`s, do
it for <100 rows. This goes along heavily with the `immutableToJs`
stuff mentioned above.
- Simplify `placeTile` a lot. Still pretty crappy but it’s way more
performant. Will fix other issues later.
- Add a few little utilities here and there.
- Vastly simplify `parseLevel` since we’re using a 2d array instead of
a stupid object.